// api-gql/src/schema/resolver-types.gen.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

export type AuthLogin = {
  __typename?: 'AuthLogin';
  firstname?: Maybe<Scalars['String']['output']>;
  refreshToken?: Maybe<Scalars['String']['output']>;
  token?: Maybe<Scalars['String']['output']>;
};

export type AuthLoginInput = {
  email: Scalars['String']['input'];
  pass: Scalars['String']['input'];
};

export type AuthSignupInput = {
  email: Scalars['String']['input'];
  firstname: Scalars['String']['input'];
  lastname: Scalars['String']['input'];
  pass: Scalars['String']['input'];
};

export type Mutation = {
  __typename?: 'Mutation';
  authLogin?: Maybe<AuthLogin>;
  authSignup?: Maybe<Result>;
};


export type MutationAuthLoginArgs = {
  input?: InputMaybe<AuthLoginInput>;
};


export type MutationAuthSignupArgs = {
  input?: InputMaybe<AuthSignupInput>;
};

export type Query = {
  __typename?: 'Query';
  version?: Maybe<Result>;
};

export type Result = {
  __typename?: 'Result';
  result: Scalars['String']['output'];
};

export type WithIndex<TObject> = TObject & Record<string, any>;
export type ResolversObject<TObject> = WithIndex<TObject>;

export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
  next: NextResolverFn<TResult>,
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => TResult | Promise<TResult>;



/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = ResolversObject<{
  AuthLogin: ResolverTypeWrapper<AuthLogin>;
  AuthLoginInput: AuthLoginInput;
  AuthSignupInput: AuthSignupInput;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']['output']>;
  Mutation: ResolverTypeWrapper<{}>;
  Query: ResolverTypeWrapper<{}>;
  Result: ResolverTypeWrapper<Result>;
  String: ResolverTypeWrapper<Scalars['String']['output']>;
}>;

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = ResolversObject<{
  AuthLogin: AuthLogin;
  AuthLoginInput: AuthLoginInput;
  AuthSignupInput: AuthSignupInput;
  Boolean: Scalars['Boolean']['output'];
  Mutation: {};
  Query: {};
  Result: Result;
  String: Scalars['String']['output'];
}>;

export type AuthLoginResolvers<ContextType = any, ParentType extends ResolversParentTypes['AuthLogin'] = ResolversParentTypes['AuthLogin']> = ResolversObject<{
  firstname?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  refreshToken?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  token?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type MutationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = ResolversObject<{
  authLogin?: Resolver<Maybe<ResolversTypes['AuthLogin']>, ParentType, ContextType, Partial<MutationAuthLoginArgs>>;
  authSignup?: Resolver<Maybe<ResolversTypes['Result']>, ParentType, ContextType, Partial<MutationAuthSignupArgs>>;
}>;

export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = ResolversObject<{
  version?: Resolver<Maybe<ResolversTypes['Result']>, ParentType, ContextType>;
}>;

export type ResultResolvers<ContextType = any, ParentType extends ResolversParentTypes['Result'] = ResolversParentTypes['Result']> = ResolversObject<{
  result?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type Resolvers<ContextType = any> = ResolversObject<{
  AuthLogin?: AuthLoginResolvers<ContextType>;
  Mutation?: MutationResolvers<ContextType>;
  Query?: QueryResolvers<ContextType>;
  Result?: ResultResolvers<ContextType>;
}>;